+++ src/components/Main/Reports/Documents/Document/Page/Page.jsx
@@ -1,5 +1,10 @@
 import $Page from "./Page.styles.jsx";
+import Table from "./Table/Table.jsx";
 
-export default function Page({ indicatorData }) {
-    return (<$Page>{JSON.stringify(indicatorData)}</$Page>);
+export default function Page({ indicatorData, section }) {
+  return (
+    <$Page>
+      <Table indicatorData={indicatorData} section={section}/>
+    </$Page>
+  );
 }